Move bitwise logical tests to separate file
