+++ tests/core/arithmetic_test.py
@@ -135,81 +135,6 @@ def test_pow_variable():
     assert np.array_equal(c.values, 2.0**data)
 
 
-def test_bitwise_ior_variable_with_variable():
-    a = sc.scalar(False)
-    b = sc.scalar(True)
-    a |= b
-    assert sc.identical(a, sc.scalar(True))
-
-    a = sc.Variable(dims=['x'], values=np.array([False, True, False, True]))
-    b = sc.Variable(dims=['x'], values=np.array([False, False, True, True]))
-    a |= b
-    assert sc.identical(
-        a, sc.Variable(dims=['x'], values=np.array([False, True, True, True])))
-
-
-def test_bitwise_or_variable_with_variable():
-    a = sc.scalar(False)
-    b = sc.scalar(True)
-    assert sc.identical((a | b), sc.scalar(True))
-
-    a = sc.Variable(dims=['x'], values=np.array([False, True, False, True]))
-    b = sc.Variable(dims=['x'], values=np.array([False, False, True, True]))
-    assert sc.identical((a | b),
-                        sc.Variable(dims=['x'],
-                                    values=np.array([False, True, True, True])))
-
-
-def test_bitwise_iand_variable_with_variable():
-    a = sc.scalar(False)
-    b = sc.scalar(True)
-    a &= b
-    assert sc.identical(a, sc.scalar(False))
-
-    a = sc.Variable(dims=['x'], values=np.array([False, True, False, True]))
-    b = sc.Variable(dims=['x'], values=np.array([False, False, True, True]))
-    a &= b
-    assert sc.identical(
-        a, sc.Variable(dims=['x'], values=np.array([False, False, False, True])))
-
-
-def test_bitwise_and_variable_with_variable():
-    a = sc.scalar(False)
-    b = sc.scalar(True)
-    assert sc.identical((a & b), sc.scalar(False))
-
-    a = sc.Variable(dims=['x'], values=np.array([False, True, False, True]))
-    b = sc.Variable(dims=['x'], values=np.array([False, False, True, True]))
-    assert sc.identical((a & b),
-                        sc.Variable(dims=['x'],
-                                    values=np.array([False, False, False, True])))
-
-
-def test_ixor_variable_with_variable():
-    a = sc.scalar(False)
-    b = sc.scalar(True)
-    a ^= b
-    assert sc.identical(a, sc.scalar(True))
-
-    a = sc.Variable(dims=['x'], values=np.array([False, True, False, True]))
-    b = sc.Variable(dims=['x'], values=np.array([False, False, True, True]))
-    a ^= b
-    assert sc.identical(
-        a, sc.Variable(dims=['x'], values=np.array([False, True, True, False])))
-
-
-def test_xor_variable_with_variable():
-    a = sc.scalar(False)
-    b = sc.scalar(True)
-    assert sc.identical((a ^ b), sc.scalar(True))
-
-    a = sc.Variable(dims=['x'], values=np.array([False, True, False, True]))
-    b = sc.Variable(dims=['x'], values=np.array([False, False, True, True]))
-    assert sc.identical((a ^ b),
-                        sc.Variable(dims=['x'],
-                                    values=np.array([False, True, True, False])))
-
-
 def test_iadd_variable_with_scalar():
     v = sc.Variable(dims=['x'], values=[10.0])
     expected = sc.Variable(dims=['x'], values=[12.0])
